generated_projects/discord_bot/requirements.txt
discord.py==2.0.0
google-cloud-sdk==365.0.0